line_bot_ai/requirements.txt
fastapi
uvicorn
requests
python-dotenv
line-bot-sdk
zhipuai
